engines/mmap: fix full/limited prep logic
The engine supports a "full" mode where the entire range for a job is mapped,
and a "limited" mode that tries to map less memory. This fixes a bug and
implements the following simplified logic:

- If we are on a 64-bit architecture, or if io_size <= mmap_map_size, then we
  map the entire io_size region, and re-use for all IO.
- Otherwise, we map the range for each individual IOs.

Previously, we would fallback to limited mode in case of 32-bit overflow of the
job size. That isn't strict enough as we want to stay below mmap_map_size for a
single mapping. This fixes that.

Also, in limited mode, this code previously tried to map min(io_size,
mmap_map_size) bytes, starting at the offset. This could allow a future IO use
the same mapping if happens at a higher offset. I think it's better to keep it
simple and get the consistent performance of always having one mmap per IO in
limited mode.